+++ Snabble/Core/API/APIVersion.swift
@@ -5,5 +5,5 @@
 //
 
 enum APIVersion {
-    static let version = "0.12.11"
+    static let version = "0.12.13"
 }